Add cluster tests and fix bugs.

+++ server/src/main/java/org/elasticsearch/readiness/ReadinessService.java
@@ -170,6 +170,9 @@ public class ReadinessService extends AbstractLifecycleComponent implements Clus
 
     @Override
     protected void doClose() throws IOException {
+        if (enabled == false) {
+            return;
+        }
         try {
             Files.deleteIfExists(getSocketPath());
         } catch (IOException e) {
